dtype fix

+++ bayesianquilts/nn/dense.py
@@ -210,7 +210,7 @@ class DenseHorseshoe(Dense, BayesianModel):
         self.bijectors = bijectors
         self.prior_distribution = tfd.JointDistributionNamed(distribution_dict)
         self.surrogate_distribution = build_surrogate_posterior(
-            self.prior_distribution, bijectors)
+            self.prior_distribution, bijectors, dtype=self.dtype)
         self.var_list = list(self.surrogate_distribution.model.keys())
 
     def sample(self, *args, **kwargs):
